Part 2: Move ship save events, remove useless/vesitgal events,  to Triad namespace, remove ShipGridData
