+++ Content.Server/Shuttles/ShipSerializationSystem.cs
@@ -12,7 +12,7 @@ using System.Collections.Generic;
 using System.Numerics;
 using YamlDotNet.Serialization;
 using YamlDotNet.Serialization.NamingConventions;
-using Content.Shared._NF.Shuttles.Save;
+using Content.Shared._Triad.Shipyard.Save;
 using Robust.Shared.Network;
 using Robust.Shared.Maths;
 using System;
