+++ Content.Shared/_Triad/Shipyard/Save/DeleteLocalShipFileMessage.cs
@@ -1,0 +1,14 @@
+using Robust.Shared.Serialization;
+
+namespace Content.Shared._Triad.Shipyard.Save;
+
+[Serializable, NetSerializable]
+public sealed class DeleteLocalShipFileMessage : EntityEventArgs
+{
+    public string FilePath { get; }
+
+    public DeleteLocalShipFileMessage(string filePath)
+    {
+        FilePath = filePath;
+    }
+}
